restructing files (removed files out of idrp-sd into parent folder) started creating outline for disease filter & state color description legend
