--- idrp-sd/src/component/diseaseData/DataView.js
@@ -1,7 +1,0 @@
-import React from "react";
-import "./DataView.css";
-import "../global.css";
-
-export default function DataView() {
-  return <div className="data-section">DataView</div>;
-}
