--- idrp-sd/src/component/global.css
@@ -1,24 +1,0 @@
-.main-container{
-display: flex;
-flex-flow: row;
-}
-
-.data-section {
-flex: 0 1 30%;
-}
-
-#map-svg{
-flex: 1 1 auto;
-height: 100%;
-overflow: hidden;
-}
-
-.single-state {
-    flex: 1 1 auto;
-}
-
-
-#HI:hover {
-
-    transform: scale(1.3);
-}
