--- idrp-sd/src/component/map/Map.css
@@ -1,1 +1,0 @@
-
