+++ src/App.css
@@ -1,0 +1,5 @@
+.main-container{
+    display: flex;
+
+    }
+
