+++ src/component/diseaseData/DataView.css
@@ -1,3 +1,4 @@
 .data-section {
     background-color: rgb(173, 144, 201);
+    flex: 0 1 25%;
 }
