+++ src/component/diseaseData/DataView.js
@@ -1,0 +1,11 @@
+import React from "react";
+import "./DataView.css";
+
+export default function DataView() {
+  return (
+    <>
+      {" "}
+      <div className="data-section">DataView</div>{" "}
+    </>
+  );
+}
