+++ src/component/diseaseData/DiseaseFilter.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function DiseaseFilter() {
+  return <div>DiseaseFilter</div>;
+}
