+++ src/component/map/ColorLegend.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function ColorLegend() {
+  return <div>ColorLegend</div>;
+}
