+++ src/component/map/Map.css
@@ -1,0 +1,4 @@
+#map-svg{
+    flex: 1 1 auto;
+    }
+
