+++ src/component/map/UsMap.css
@@ -1,0 +1,9 @@
+.svg-container{
+    width: 100%;
+    height: 100%;
+}
+
+/* svg {
+    width: 100%;
+    height: 90%;
+} */
